refactor(rig): rig_probe.py takes the bpy rig probing out of request_builder
armature_to_skeleton, is_control_rig, deform-bone detection, the action
frame range and user-edit detection are Blender-only and stay in the
addon; request_builder keeps them as a re-export until it is replaced by
the shared core builder.
